ChibiOS: fix deadlocks when hammering mouse/extra keys
When hammering on the mouse or extra keys, tmk will deadlock if ChibiOS was
compiled with assertions. This particular assertion in usbTransmitI fires
if the last packet hasn't been acknowledged yet by the host driver.

To fix that, we'll save the new report and transmit it if the last
report was acknowledged.